02.09.2021 папки синхронизируются. баг- во время синхронизации(во время пересылки файлов) папки не отслеживаются, чтобы не пладить лишние потоки синхронихации.  можно поставить задержку в вотчере на несколько минут. а лучще в конце синхронизации сравнить каталоги.

+++ serverPart/src/main/java/service/FileProcessing.java
@@ -9,21 +9,19 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
-import java.util.ArrayList;
 import java.util.List;
 
 public class FileProcessing {
 
 
-
     public static void saveAsFile(ObjectCreatorClass filePrepare) {
 
         try {
-            File fileА = new File(String.valueOf("SERVER_FOLDER/" + filePrepare.getFileName()));
-            FileOutputStream fos = new FileOutputStream(fileА);
+            File fileA = new File(String.valueOf("SERVER_FOLDER/" + filePrepare.getFileName()));
+            FileOutputStream fos = new FileOutputStream(fileA);
             fos.write(filePrepare.getFileBin());//записали
             //исправляем дату послднего измененя (ставим как у клиента)
-            fileА.setLastModified(filePrepare.getLastModify());
+            fileA.setLastModified(filePrepare.getLastModify());
             //    System.out.println(filePrepare.getLastModify());
             fos.close();
         } catch (IOException e) {
@@ -34,19 +32,19 @@ public class FileProcessing {
     }
 
     public static void deleteFile(List<FileProperty> fileList) {
+        if (fileList.size() > 0) {
+            for (FileProperty file : fileList) {
+                Path path = Paths.get(file.getPath());
+                System.out.println(file.getPath() + " deleted!");
+                try {
+                    Files.delete(path);
+                } catch (IOException e) {
+                    e.printStackTrace();
+                }
 
-        for (FileProperty file : fileList) {
-            Path path = Paths.get(file.getPath());
-            System.out.println(file.getPath()+" deleted!");
-            try {
-                Files.delete(path);
-            } catch (IOException e) {
-                e.printStackTrace();
             }
-
+            System.out.println("Files deleted!");
         }
-        System.out.println("Files deleted!");
     }
-
 
 }
